completed client, portfolio, report and restructured project

+++ src/main/java/com/smartstake/api/model/repositories/PortfolioRepository.java
@@ -3,6 +3,6 @@ package com.smartstake.api.model.repositories;
 import com.smartstake.api.model.entities.*;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PortfolioRepository extends JpaRepository<PortfolioEntity, Integer> {
+public interface PortfolioRepository extends JpaRepository<PortfolioEntity, Long> {
 
 }
